backend/requirements.txt
flask==3.1.3
gunicorn
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
rq==2.7.0
redis==7.2.0
psycopg2-binary
alembic